requirements.txt
azure-identity>=1.16.0
azure-storage-blob>=12.20.0
openai>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
streamlit>=1.40.0
pymupdf>=1.24.0
pandas>=2.0.0
